Обновил пермишены
Сделал норм генерацию логина и пароля
