+++ CRM/rules.py
@@ -38,3 +38,4 @@ rules.add_perm('worker', is_logged_admin)
 
 rules.add_perm('profile', is_logged_personnel)
 rules.add_perm('add_user', is_logged_manager_admin)
+rules.add_perm('users_list', is_logged_personnel)
