+++ CRM/templates/CRM/users/password.html
@@ -17,10 +17,17 @@
             {% csrf_token %}
             <div class="form-group">
                 {% if password %}
-                    {{ password }} <p></p>
+                    Логин: {{ username }}<p></p>
+                    Пароль: {{ password }}<p></p>
                 {% endif %}
             </div>
         </form>
+        <a name="btn-success"
+             href="{% url 'users:list' %}"
+             class="btn btn-sm btn-success"
+          >
+            Ок
+          </a>
     </div>
 </div>
 {% endblock content %}
