+++ CRM/templates/CRM/users/users.html
@@ -11,8 +11,17 @@
 
 {% block content %}
   <div class="content-section ml-4 mt-4 w-50">
-    <legend class="border-bottom mb-2">Список работников</legend>
+      <legend class="border-bottom mb-2">Список работников </legend>
     <form method="post">
+        {% if not user.is_worker_user %}
+        <a name="add-new"
+             href="{% url 'users:add' %}"
+             class="btn btn-sm btn-info"
+          >
+            Добавить сотрудника
+          </a>
+        <p></p>
+        {% endif %}
        <table  border="1">
            <tr><td width="300"><b>ФИО</b></td>
           <td width="300"><b>Должность</b></td>
